+++ frontend/src/app/app-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { RouteListViewComponent } from './route/list/route-list-view.component';
-import { TripViewComponent } from './trip/trip-view.component';
-import { RouteSelectionViewComponent } from './route-selection/route-selection-view.component';
+import { RouteListViewComponent } from './views/route-list/route-list-view.component';
+import { TripViewComponent } from './views/trip/trip-view.component';
+import { RouteSelectionViewComponent } from './views/route-selection/route-selection-view.component';
 
 const routes: Routes = [
     { path: 'newTrip', component: TripViewComponent },
