+++ frontend/src/app/app.module.ts
@@ -7,13 +7,13 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
-import { MainMenuComponent } from './shared/menu/main-menu.component';
-import { RouteListViewComponent } from './route/list/route-list-view.component';
-import { RouteService } from './route/route.service';
-import { TripViewComponent } from './trip/trip-view.component';
-import { TripService } from './trip/trip.service';
-import { TimePipe } from './shared/time.pipe';
-import { RouteSelectionViewComponent } from './route-selection/route-selection-view.component';
+import { MainMenuComponent } from './components/main-menu/main-menu.component';
+import { RouteListViewComponent } from './views/route-list/route-list-view.component';
+import { RouteService } from './services/route.service';
+import { TripViewComponent } from './views/trip/trip-view.component';
+import { TripService } from './services/trip.service';
+import { TimePipe } from './pipes/time.pipe';
+import { RouteSelectionViewComponent } from './views/route-selection/route-selection-view.component';
 
 @NgModule({
     declarations: [
