+++ frontend/src/app/domain/trip.ts
