+++ frontend/src/app/pipes/time.pipe.ts
