+++ frontend/src/app/views/trip/trip-view.component.html
